fix change

+++ testJWT/internal/service/article_service.go
@@ -15,15 +15,15 @@ func NewArticleService(repo *repository.ArticleRepository) *ArticleService {
 	return &ArticleService{repo: repo}
 }
 
-func (s *ArticleService) CreateArticle(ctx context.Context, authorID int64, input model.ArticleInput) (*model.Article, error) {
+func (s *ArticleService) CreateArticle(ctx context.Context, authorID uint, input model.ArticleInput) (*model.Article, error) {
 	return s.repo.CreateArticle(ctx, authorID, input.Title, input.Content)
 }
 
-func (s *ArticleService) GetArticle(ctx context.Context, id int64) (*model.Article, error) {
+func (s *ArticleService) GetArticle(ctx context.Context, id uint) (*model.Article, error) {
 	return s.repo.GetArticleByID(ctx, id)
 }
 
-func (s *ArticleService) GetArticlesByAuthor(ctx context.Context, authorID int64) ([]*model.Article, error) {
+func (s *ArticleService) GetArticlesByAuthor(ctx context.Context, authorID uint) ([]*model.Article, error) {
 	return s.repo.GetArticlesByAuthor(ctx, authorID)
 }
 
@@ -31,6 +31,6 @@ func (s *ArticleService) UpdateArticle(ctx context.Context, id int64, input mode
 	return s.repo.UpdateArticle(ctx, id, input.Title, input.Content)
 }
 
-func (s *ArticleService) DeleteArticle(ctx context.Context, id int64) error {
+func (s *ArticleService) DeleteArticle(ctx context.Context, id uint) error {
 	return s.repo.DeleteArticle(ctx, id)
 }
